Fixed BFS implementation for PDB generation:
Previous implementation ignored fact that some edges had 0-weight. This meant that nodes could be marked as visited when they weren't the lowest-weight nodes (preventing future visits with potentially lower weights). To fix this, 0-weight edge nodes were added to the beginning of the open set and 1-weight nodes to the end of the open set (while still popping from the beginning), ensuring that nodes are popped in non-decreasing weight order. This is only possible because the graph is binary-weighted (edges are only 0 and 1 cost). Otherwise, Djikstra would be necessary.